Switch salt-master metaparameter from 'notify' to 'before'
This is a weird behavior I've seen with Puppet where 'notify' doesn't
alway handle the current state of services as expected. Ideally the
salt-master daemon should be notified to restart whenever there are
config file changes, however that doesn't work with the initial setup in
this case, so we're stuck using 'before' instead.