Use salt provisioner.
Since this project is about learning salt, we use the vagrant salt
provisioner to bring up the salt parts of the nodes. In particular,
all the puppet manifests are removed so that this project only depends
on salt-centric resources.

We add a couple of simple Makefiles to facilitate creating the keys
needed by the salt master and the minions. It's not necessary to use
the Makefiles -- but the keys themselves are required. This way, we
don't have to distribute a set of keys with the salt-sandbox
repository.

We copy the included "hello, world" salt state into a directory that
is shared with the vagrant guest running the salt master. The state
tree is ignored by this git repository so anyone can start
experimenting with salt without having to worry about this repository
trying to track their work.